requirements.txt
flask==3.0.3
flask-restful==0.3.10
flask-jwt-extended==4.6.0
Flask-PyMongo==2.3.0
bcrypt==4.1.3
python-dotenv==1.0.1
pycryptodome==3.21.0
Werkzeug==3.0.4
pymongo==4.10.1
